Remove kubernetes
- Learning kubernetes was worthwhile, and I would still like to return
to using it in the future, but I haven't tested the kubernetes setup in
quite some time and it is likely out of date.  With the switch to a
django backend, the inconsistencies will likely pile up, so I am
removing for now, but will add back in future once things are clearer.
For now, docker compose will rule the kingdom.