Improve docker ignore
- Previously was assuming paths in dockerignore applied for
subdirectories, but turns out paths provided are relative to directory
that the .dockerignore file is located in.  Because of the setup of how
client container is built (context is root dir instead of client/), I
have updated some of the paths to properly ignore some files inside
client or elsewhere.